Revert "notwork"
This reverts commit 14c35a2e3e134c77311437dcc30e1987cbcb0621.

	изменено:      SICK_Program/Program.cs
	изменено:      SICK_Program/WebAPIConfigTypes.cs
	изменено:      SICK_Program/WebConfig.json
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/asset-manifest.json
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/icon.ico
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/manifest.json
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/css/main.e0f226ea.css
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/css/main.e0f226ea.css.map
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/js/787.ec704a7a.chunk.js
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/js/787.ec704a7a.chunk.js.map
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/js/main.4dc7c3da.js
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/js/main.4dc7c3da.js.map
	удалено:       SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/media/logo_dark.55e6f5eb7c8c5f1d982e0a5335a35454.svg
	изменено:      SICK_Program/wwwroot/asset-manifest.json
	новый файл:    SICK_Program/wwwroot/css/site.css
	изменено:      SICK_Program/wwwroot/index.html
	переименовано: SICK_Program/wwwroot/2023-07-14-17-01-56-111/index.html -> SICK_Program/wwwroot/index.html.bak
	новый файл:    SICK_Program/wwwroot/index_.html
	новый файл:    SICK_Program/wwwroot/js/script.js
	удалено:       SICK_Program/wwwroot/static/js/main.4dc7c3da.js
	удалено:       SICK_Program/wwwroot/static/js/main.4dc7c3da.js.LICENSE.txt
	удалено:       SICK_Program/wwwroot/static/js/main.4dc7c3da.js.map
	новый файл:    SICK_Program/wwwroot/static/js/main.ceb38239.js
	переименовано: SICK_Program/wwwroot/2023-07-14-17-01-56-111/static/js/main.4dc7c3da.js.LICENSE.txt -> SICK_Program/wwwroot/static/js/main.ceb38239.js.LICENSE.txt
	новый файл:    SICK_Program/wwwroot/static/js/main.ceb38239.js.map
	изменено:      Sick-scanner/Class1.cs
	изменено:      Sick-scanner/Filter.cs
	изменено:      Sick-scanner/config.cs

+++ Sick-scanner/config.cs
@@ -83,9 +83,9 @@
         ///<summary>частота сканирования</summary>
         public int Frequency { get; set; }
         ///<summary>начальный угол</summary>
-        public double StartAngle { get; set; }
+        public int StartAngle { get; set; }
         ///<summary>конечный угол</summary>
-        public double EndAngle { get; set; }
+        public int EndAngle { get; set; }
         ///<summary>единичный угол отклонения(шаг)</summary>
         public double Resolution { get; set; }
     }
